08-kafka-pubsub-streaming/python-client/requirements.txt
kafka-python
Faker